+++ Changelog.md
@@ -9,6 +9,7 @@
 - Fix itemgiver: negative keep # wasn't properly tracking items already given
 - Add QuestTracker.AutoRequest setting
 - Rewrite DungeonMaps with performance in mind
+- Fix equip mgr command pattern, default profile selection, and clean up event handling
 
 ## 0.1.1 (2019-12-03) [UtilityBeltInstaller-0.1.1.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/c94b459edb29649f99f1d84eae3f76ce/UtilityBeltInstaller-0.1.1.exe)
 - Increase AutoStack/AutoCram timeout to 30s